Move entity commands into core module
