+++ src/command/index.js
@@ -1,5 +1,4 @@
 export * from './plugin.js'
 export * from './core/index.js'
-export * from './commands/index.js'
 export * from './resources/index.js'
 export * from './typedef/index.js'
