+++ src/command/resources/index.js
@@ -1,2 +1,1 @@
 export * from './queue.js'
-export * from './entity.js'
