+++ src/core/core/index.js
@@ -1,2 +1,3 @@
 export * from './schedules.js'
 export * from './runner.js'
+export * from './entity.js'
