+++ src/core/index.js
@@ -1,2 +1,3 @@
+export * from './commands/index.js'
 export * from './core/index.js'
 export * from './plugin.js'
